dtmq的wal_handle部分迁移完成
^ Conflicts:
^	src/component/dtmq/dtmq-proxysvr/CMakeLists.txt

+++ src/component/dtmq/sdk/common/rpc/dtmq/dtmq_algorithm.cpp
@@ -6,15 +6,15 @@
 
 #include <string/string_format.h>
 
-//clang-format off
+// clang-format off
 #include <config/compiler/protobuf_prefix.h>
-//clang-format on
+// clang-format on
 
 #include <protocol/pbdesc/com.struct.dtmq.pb.h>
 
-//clang-format off
+// clang-format off
 #include <config/compiler/protobuf_suffix.h>
-//clang-format on
+// clang-format on
 
 #include <cstdint>
 #include <string>
@@ -22,19 +22,19 @@
 namespace rpc {
 namespace dtmq {
 
-DTMQ_COMMON_SDK_API size_t get_hash_code(const atfw::dtmq::DChannelMessage& channel_log) noexcept {
+DTMQ_COMMON_SDK_API uint64_t get_hash_code(const atfw::dtmq::DChannelMessage& channel_log) noexcept {
   return channel_log.hash_code();
 }
 
-DTMQ_COMMON_SDK_API void set_hash_code(atfw::dtmq::DChannelMessage& channel_log, size_t hash_code) noexcept {
+DTMQ_COMMON_SDK_API void set_hash_code(atfw::dtmq::DChannelMessage& channel_log, uint64_t hash_code) noexcept {
   channel_log.set_hash_code(hash_code);
 }
 
-DTMQ_COMMON_SDK_API size_t calculate_hash_code(size_t previous,
-                                               const atfw::dtmq::DChannelMessage& channel_log) noexcept {
-  size_t buffer[1] = {static_cast<size_t>(channel_log.sequence())};
+DTMQ_COMMON_SDK_API uint64_t calculate_hash_code(uint64_t previous,
+                                                 const atfw::dtmq::DChannelMessage& channel_log) noexcept {
+  uint64_t buffer[1] = {static_cast<uint64_t>(channel_log.sequence())};
 
-  return static_cast<size_t>(XXH64(buffer, sizeof(buffer), static_cast<XXH64_hash_t>(previous)));
+  return static_cast<uint64_t>(XXH64(buffer, sizeof(buffer), static_cast<XXH64_hash_t>(previous)));
 }
 
 DTMQ_COMMON_SDK_API std::string make_unicast_channel_id(uint32_t type_id, uint64_t zone_id, uint64_t instance_id) {
